Fixed use of wrong loop, added 6 tests for dealCards and some refactoring.
Wrong foor loop: Replaced for of with a traditional for loop as the for of value only existed inside the loop's scope, which caused the modified values to never get carried back out. This made the variable pretty much useless so it was better to just use a good ol' for loop.

Added 6 unit tests:
 - "dealCards - Modifies supplied decks."
 - "dealCards - All source card decks have been dealt."
 - "dealCards - Dealt decks consist of cards from the source card deck."
 - "dealCards - Dealt decks have equal length when even amount."
 - "dealCards - Dealt decks have equal length when oddamount."
 - "dealCards - Dealt decks have no card in common."